seperating route validation if only so it makes the integration testing a little less suck
