+++ src/MGNZ.Squidex.CLI.Common/Routing/StaticCommandLineRouter.cs
@@ -1,8 +1,6 @@
 namespace MGNZ.Squidex.CLI.Common.Routing
 {
-  using System;
   using System.Collections.Generic;
-  using System.Linq;
 
   using MGNZ.Squidex.CLI.Common.Commands;
 
@@ -10,17 +8,20 @@ namespace MGNZ.Squidex.CLI.Common.Routing
 
   public class StaticCommandLineRouter : ICommandLineRouter
   {
-    private readonly ILogger _logger;
     private readonly IParseRouteCommands _commandParser;
+    private readonly ILogger _logger;
     private readonly IParseRouteOptions _optionParser;
+    private readonly IValidateRouteRequests _requestValidator;
     private readonly IBuildRouteRequests _routeBuilder;
 
-    public StaticCommandLineRouter(ILogger logger, IParseRouteCommands commandParser, IParseRouteOptions optionParser, IBuildRouteRequests routeBuilder)
+    public StaticCommandLineRouter(ILogger logger, IParseRouteCommands commandParser, IParseRouteOptions optionParser,
+      IBuildRouteRequests routeBuilder, IValidateRouteRequests requestValidator)
     {
       _logger = logger;
       _commandParser = commandParser;
       _optionParser = optionParser;
       _routeBuilder = routeBuilder;
+      _requestValidator = requestValidator;
     }
 
     public BaseRequest GetOne(IEnumerable<Noun> cachedNouns, string[ ] arguments)
@@ -29,16 +30,7 @@ namespace MGNZ.Squidex.CLI.Common.Routing
       _optionParser.ParseAndPopulateOptions(ref nounVerbCombination, arguments);
       var request = _routeBuilder.GetRequestForVerb(nounVerbCombination);
 
-      var validity = request.Validate();
-      if (validity.Any(v => v.isValid == false))
-      {
-        var reasons = validity.TakeWhile(f => f.isValid == false).Select(s => $"- {s.property}: {s.invalidReason}").ToList();
-        var message = $"The request to route the command '{string.Join(" ", arguments)}' into a {request.GetType().Name} failed because due to the following reasons{Environment.NewLine}{string.Join(Environment.NewLine,reasons)}";
-        var error = new ArgumentException(message);
-
-        _logger.Error(error, "The request to route the {@arguments} into a {@request} failed due to the following {@reasons}", arguments, request, reasons);
-        throw error;
-      }
+      _requestValidator.Validate(arguments, request);
 
       return request;
     }
@@ -56,6 +48,5 @@ namespace MGNZ.Squidex.CLI.Common.Routing
 
   public interface ICommandLineRouter
   {
-
   }
 }
